deploying database please

+++ backend/CineNiche.API/Data/MoviesUser.cs
@@ -16,7 +16,7 @@ public partial class MoviesUser
 
     public string? Email { get; set; }
 
-    public int? Age { get; set; }
+    public byte? Age { get; set; }
 
     public string? Gender { get; set; }
